requirements.txt
langchain==0.1.16
openai>=1.0.0
python-dotenv==1.0.1
tiktoken>=0.5.0
langchain-groq
langchain-google-community[texttospeech]
pygame>=2.1.0
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
pytest>=7.0.0
